# apps/platform-api/specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API Documentation",
    "version": "1.0.0",
    "description": "API Documentation"
  },
  "paths": {
    "/api/v1/admin/user-permissions/grant": {
      "post": {
        "summary": "为用户授予权限",
        "tags": [
          "Admin",
          "Permission"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "minLength": 36,
                    "maxLength": 36
                  },
                  "permissionIds": {
                    "type": "array",
                    "items": {
                      "type": "number",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  }
                },
                "required": [
                  "userId",
                  "permissionIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "权限授予成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "granted": {
                      "type": "number"
                    },
                    "skipped": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "granted",
                    "skipped"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/user-permissions/revoke": {
      "post": {
        "summary": "撤销用户权限",
        "tags": [
          "Admin",
          "Permission"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "minLength": 36,
                    "maxLength": 36
                  },
                  "permissionIds": {
                    "type": "array",
                    "items": {
                      "type": "number",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  }
                },
                "required": [
                  "userId",
                  "permissionIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "权限撤销成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revoked": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "revoked"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/audit": {
      "get": {
        "summary": "列出所有审计日志",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "审计日志列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "actor_id": {
                            "type": "string"
                          },
                          "actor_role": {
                            "type": "object"
                          },
                          "action": {
                            "type": "string"
                          },
                          "target_type": {
                            "type": "string"
                          },
                          "target_id": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "SUCCESS",
                              "FAILURE"
                            ]
                          },
                          "message": {
                            "type": "object"
                          },
                          "meta": {
                            "type": "object"
                          },
                          "created_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "actor_id",
                          "actor_role",
                          "action",
                          "target_type",
                          "target_id",
                          "status",
                          "message",
                          "meta",
                          "created_at"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit/{id}": {
      "get": {
        "summary": "根据ID获取审计日志",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "审计日志详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "actor_id": {
                      "type": "string"
                    },
                    "actor_role": {
                      "type": "object"
                    },
                    "action": {
                      "type": "string"
                    },
                    "target_type": {
                      "type": "string"
                    },
                    "target_id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "SUCCESS",
                        "FAILURE"
                      ]
                    },
                    "message": {
                      "type": "object"
                    },
                    "meta": {
                      "type": "object"
                    },
                    "created_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "actor_id",
                    "actor_role",
                    "action",
                    "target_type",
                    "target_id",
                    "status",
                    "message",
                    "meta",
                    "created_at"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}